Decode multiple metadata versions
Tests pass
