+++ core/Cargo.toml
@@ -7,8 +7,10 @@ edition = "2018"
 [dependencies]
 failure = "0.1.6"
 codec = { version = "1.1.0", package = "parity-scale-codec" }
+codec411 = { version = "4.1.1", package = "parity-codec", features = ["derive"] } # for backwards compatibility with v 7
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0" }
+
 runtime-metadata-latest = { git = "https://github.com/paritytech/substrate/", package = "frame-metadata", branch = "polkadot-master" }
 runtime-metadata09 = { git = "https://github.com/paritytech/substrate/", package = "frame-metadata", rev = "a512dcdd3931aa571c92c8550a0e079d5b27239a" }
 runtime-metadata08 = { git = "https://github.com/paritytech/substrate/", package = "srml-metadata", rev = "ac11c330a95827211e1999781b5f691e72387c9e" }
